SQL code to extract data
This code is joining selected variables to a cohort of patients with clinical depression. These patients have an active referral within a six month time frame of 1st of March 2015 to 31st of August 2015. Each variable selected is extracted to reflect patient behaviour 12 months before the time frame (1st of March 2014 to 30th April 2015) and during the time frame as well. This repo is a documentation of the changes and updates made to it from the 18/07/2016 onwards.